Split add-in settings dialog and extract shared RadioCard primitive
AddinSettingsDialog drops from 575 LoC to a 216-line shell. Per-tab
content moves to UserSettingsTabContent (form state + save) and
BehaviorTabContent (Outlook session-policy radios), making the three
tabs symmetric with the existing AppearanceTabContent.

RadioCard converges the two divergent radio-as-card implementations
(AppearanceTabContent's <button role="radio"> and the dialog's
<label>+<input>) onto a single <label>+<input type="radio"> primitive.
Native arrow-key navigation now works in the appearance dialog.
